# [bluetooth_proxy] Remove ESPBTUUID dependency to save 296 bytes of flash