--- a/esphome/components/ble_client/__init__.py
+++ b/esphome/components/ble_client/__init__.py
@@ -285,6 +285,7 @@ async def remove_bond_to_code(config, action_id, template_arg, args):
 
 async def to_code(config):
     # Register the loggers this component needs
+    cg.add_define("USE_ESP32_BLE_UUID")
     esp32_ble.register_bt_logger(BTLoggers.GATT, BTLoggers.SMP)
 
     var = cg.new_Pvariable(config[CONF_ID])