--- a/esphome/components/esp32_ble/ble_uuid.cpp
+++ b/esphome/components/esp32_ble/ble_uuid.cpp
@@ -1,7 +1,7 @@
 #include "ble_uuid.h"
 
 #ifdef USE_ESP32
-#ifdef USE_ESP32_BLE_DEVICE
+#ifdef USE_ESP32_BLE_UUID
 
 #include <cstring>
 #include <cstdio>
@@ -191,5 +191,5 @@ std::string ESPBTUUID::to_string() const {
 
 }  // namespace esphome::esp32_ble
 
-#endif  // USE_ESP32_BLE_DEVICE
+#endif  // USE_ESP32_BLE_UUID
 #endif  // USE_ESP32