--- a/esphome/components/esp32_ble_tracker/esp32_ble_tracker.h
+++ b/esphome/components/esp32_ble_tracker/esp32_ble_tracker.h
@@ -33,12 +33,12 @@ enum AdvertisementParserType {
   RAW_ADVERTISEMENTS,
 };
 
+#ifdef USE_ESP32_BLE_DEVICE
 struct ServiceData {
   ESPBTUUID uuid;
   adv_data_t data;
 };
 
-#ifdef USE_ESP32_BLE_DEVICE
 class ESPBLEiBeacon {
  public:
   ESPBLEiBeacon() { memset(&this->beacon_data_, 0, sizeof(this->beacon_data_)); }