… esphome/core/defines.h