# What does this implement/fix?

This PR removes the unnecessary ESPBTUUID dependency from bluetooth_proxy's `fill_128bit_uuid_array` function, saving 296 bytes of flash and improving performance during the critical GATT service discovery phase.

**The problem:**
- bluetooth_proxy was using `esp_bt_uuid_t uuid = espbt::ESPBTUUID::from_uuid(uuid_source).as_128bit().get_uuid();` 
- This was massive overkill - creating temporary objects and calling multiple methods just to pack UUID bytes
- It forced bluetooth_proxy to depend on and link the entire ESPBTUUID class
- **Critical performance impact:** During GATT service discovery, the ESP32 is under extreme pressure - fetching all services, characteristics, and descriptors while proxying them to Home Assistant. This function is called hundreds of times in this hot path.

**The solution:**
- Directly pack the UUID bytes into the output array using bit manipulation
- Use precalculated constants for the Bluetooth base UUID (0x00001000ULL and 0x800000805F9B34FBULL)
- Handle all UUID types (16/32/128-bit) with simple ternary expressions
- No temporary objects, no method calls, no ESPBTUUID dependency

**Results:**
- **Flash savings:** 296 bytes
- **Performance:** 58% faster (1.58x speedup on x86_64)
  - 16-bit UUIDs: 48% faster (1.48x speedup)
  - 32-bit UUIDs: 79% faster (1.79x speedup)  
  - 128-bit UUIDs: 20% faster (1.20x speedup)
  - **Note:** These benchmarks are from an x86_64 host system. On ESP32's slower processor, the actual time savings are ~20x larger, making this optimization even more critical during connection establishment
- **Reduced connection bottleneck:** Significantly less CPU time spent on UUID conversion during GATT discovery means more available for BLE/WiFi communication
- **Correctness:** Verified with 76,555 tests including all 65,536 possible 16-bit UUIDs

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):** N/A

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** N/A

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx
- [ ] nRF52840

## Example entry for `config.yaml`:

```yaml
# No configuration changes required - this is an internal optimization
bluetooth_proxy:
  active: true
```

## Supporting Changes

To enable bluetooth_proxy to work without ESPBTUUID:

1. **Added `USE_ESP32_BLE_UUID` define** for conditional compilation
   - Components that need ESPBTUUID (`esp32_ble_server`, `esp32_ble_tracker`, `ble_client`) add this define
   - bluetooth_proxy no longer needs to add this define

2. **Wrapped ESPBTUUID with guards** in `esp32_ble/ble_uuid.h` and `esp32_ble/ble_uuid.cpp`
   - ESPBTUUID is only compiled when `USE_ESP32_BLE_UUID` is defined
   - Added `#include "esphome/core/defines.h"` to properly check for the define

3. **Wrapped ServiceData struct** in `esp32_ble_tracker/esp32_ble_tracker.h`
   - ServiceData depends on ESPBTUUID, so it also needs the guard
   - Prevents compilation errors when ESPBTUUID is not available

## Performance Benchmark Results

Comprehensive testing comparing original vs optimized implementation:

```
=== CORRECTNESS TESTS - COMPARING ORIGINAL VS OPTIMIZED ===
Testing full range of 16-bit UUIDs (0x0000 to 0xFFFF)...
✓ All 65,536 16-bit UUIDs passed!

Testing sampled 32-bit UUIDs...
✓ All tested 32-bit UUIDs passed (10016 samples)!

Testing random 128-bit UUIDs...
✓ All 1000 random 128-bit UUIDs passed!

=== SUMMARY ===
Total tests run: 76555
Tests failed: 0

=== PERFORMANCE BENCHMARK - ORIGINAL VS OPTIMIZED ===
16-bit UUIDs:
  Original:  1.29 ns/call
  Optimized: 0.87 ns/call (speedup: 1.48x)
32-bit UUIDs:
  Original:  1.55 ns/call
  Optimized: 0.87 ns/call (speedup: 1.79x)
128-bit UUIDs:
  Original:  1.41 ns/call
  Optimized: 1.17 ns/call (speedup: 1.20x)

Overall Performance:
  Original:  1.91 ns/call
  Optimized: 1.21 ns/call (speedup: 1.58x)
```

## Technical Details

The optimization replaces runtime bit shifting with precalculated constants:
- Base UUID bytes 0-7: `0x800000805F9B34FBULL`
- Base UUID bytes 8-11: `0x00001000ULL`

The function now directly packs UUID values without intermediate buffers or object creation.

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).
  
Note: Comprehensive testing was performed with standalone test programs that verify both correctness and performance. The tests confirm all UUID types (16-bit, 32-bit, 128-bit) are handled correctly with the expected output.

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).